For CASSANDRA-9528, output junit tests to a file per suite. Redirect stdout and stderr to logback so they also end up in the file interleaved nicely. Also modify the brief formatter so when it logs test completion it also logs the suite name since they can interlave with test start messages.
For CASSANDRA-9528, capture stdout/stderr to log file, but also tee it to the console. When running all the tests don't print test output to the console, but when running a specific test do send the output to the console

For CASSANDRA-9528, make it possible to override cassandra.keepBriefBrief

For CASSANDRA-9528, fix merge issue with eclipse-warnings static analysis

For CASSANDRA-9528, bring back rolling file appender and compression because some tests are absolutely retarded about log output.

For CASSANDRA-9528, minor typo and whitespace fixes.

For CASSANDRA-9528, bring back rolling file appender and compression because some tests are absolutely retarded about log output.

For CASSANDRA-9528, output the same log directory as before