For CASSANDRA-9526, expose FailureDetector phi values via JMX and tweak trace and debug logging
 This is a combination of 4 commits.
 The first commit's message is:
Resolve unlikely race condition in phi reporting

If two phi values come in at the same time as the first one for a host, one could be lost due to the way the values are being added to the Hashtable. This code resolves that window, by switching to a ConcurrentHashMap and using putIfAbsent to atomically check for a prior value.

I doubt this could actually happen in the wild but it's still good defensive coding. Also, it removes Hashtable which is always synchronized.

 This is the 2nd commit message:

Expose the PHI values via JMX

 This is the 3rd commit message:

Start generating debug log messages when PHI gets close

This is a great way to see that phi_convict_threshold might be too low. It's not WARN or even INFO because this could generate a lot of logs, but arguably it could be. If someone has trouble with nodes going offline, they can turn up the debugging levels and see that phi_convict_threshold is the culprit.

Also contains some code cleanup.

 This is the 4th commit message:

For CASSANDRA-9526 update changes.txt