Resolve unlikely race condition in phi reporting
If two phi values come in at the same time as the first one for a host, one could be lost due to the way the values are being added to the Hashtable. This code resolves that window, by switching to a ConcurrentHashMap and using putIfAbsent to atomically check for a prior value.

I doubt this could actually happen in the wild but it's still good defensive coding. Also, it removes Hashtable which is always synchronized.