For CASSAMDRA-9462, implement correct search for SSTableIntervalTree for bounds that wrap. Implement unwrapping for all bounds types.
For CASSANDRA-9462, initial fix by Benedict.

For CASSANDRA-9462, fix the fix to get most of the tests passing. Not confident yet that this right. Implement unwrapping for all bounds types

For CASSANDRA-9462, further fixes, turns out when right is min() it might mean everything is < it as well as any such bounds being non-empty. The value of left doesn't really seem to matter.

For CASSANDRA-9462, remove unwrap implementations I added since the classes aren't supposed to be able to wrap and there have always been assertions to that effect in the ctors.

For CASSANDRA-9462, whitespace and typo fixes