For CASSANDRA-10155, stop using cfId in caches, use KS name + CF name to avoid ambiguity with 2i and base tables.
Also collapse saved cache files for different CFs (and KSs) and write them all to a single file. This eliminates file naming ambiguity.

The loop loading keys for caches was collecting an unbounded number of futures of asychronous work pending load into the cache. That is a little unwise so now it processes the work as soon as it is available or stops if a high watermark is reached.

Since caches are no longer per CF they are also not saved and loaded per CF. Saving was already not really per CF, but loading was. Loading is async so the key and row cache can load in parallel. They can also load concurrently with other things during startup, but I elected to not change existing behavior so it submits the tasks and then blocks on the result.

Counter cache is loaded by itself since it has to init after StorageProxy has done some voodoo.

In terms of negative memory impact there is no change for the on heap caching. It's still a reference to canonical object so the overhead of keys in caches haven't changed. There is additional disk utilization and encoding for saving caches. The OHC is impacted because it now has to store the KS and CF names, but this is unavoidable until UUIDs for CFs are not ambigous, or alternatively conditional encoding is done to disambiguate. Probably not worth it given they are max 48 chars each. The extra encoding/decoding is the least welcome impact.