For CASSANDRA-10688, don't bound the search space or visited set and fix some bugs and add tests.
Fix bug where array elements aren't traversed.
For Iterables, Maps, and Object[], iterate their entries instead of traversing the objects as a graph.
For a linked list this addresses the stack utilization due to the deep recursion (even if the stack is stored on the heap).
Add tests for a linked list, map, and array, and validate that roughly the expected number of objects are visited as well as the number of iterations taken it visit them.

No sure what is going on yet. CLQ doesn't traverse correctly.

Think I got it this time

Why not NBHM

Update fudge factor to reflect platform differences