Improve IntervalTree build throughput
IntervalTree does a ton of redundant sorting as it builds the tree when it could sort once and then leverage the already sorted order.

Also include a more efficient add method from CASSANDRA-20164 and periodically rebuild the tree to deal with the lack of rebalancing.

Patch by Ariel Weisberg and Yuqi Yan; Reviewed by Marcus Eriksson for CASSANDRA-19596

Co-authored-by: Yuqi Yan <yukei0509@gmail.com>