Improve IntervalTree build throughput
IntervalTree does a lot of redundant sorting as it builds the tree when it could sort once and then leverage the already sorted order.

Update `IntervalNode` constructor to preserve the sorted order of the input so it doesn't have to sort every time and use a slightly different method to calculate the midpoint using the minOrder and maxOrder lists.

Allocate appropriately sized lists in `IntervalNode` constructor to reduce unnecessary ArrayList resizing.

Add an `update` method that takes removals and additions, sorts the removals/addition lists, and then does a linear merge with the existing sorted arrays.

This version is 2x faster to construct a tree from a list and 4x faster testing with intervals that mimic leveled compaction, and produces identical trees to compared to the previous versions.

Also perform the building of the SSTableIntervalTree in Tracker under a lock instead of a CAS loop since it takes a while to build a tree and contention is expected.